Refactor unit converter to accept double values and update ReadMe.md with project details

+++ PhaseOne/1_UnitConverter.dart
@@ -1,10 +1,10 @@
 void main() {
 var kg = converter(2);
-print(kg);
+print(kg); // 2 Kg is 4.4 lbs
 }
 
 // The Converter Function
-converter(int kilo) {
+converter(double kilo) { // Edited: int type will broke the code if i added smth like 2.5
     const KgToPound = 2.2;
     return "$kilo Kg is ${kilo * KgToPound} lbs";
 }
